+++ src/Functions/Throw/throwEx.ts
@@ -1,0 +1,5 @@
+const throwEx = (msg: string) => {
+  throw Error(msg);
+};
+
+export default throwEx;
