+++ src/components/Form/FormTypes/SupervisorName.ts
@@ -1,0 +1,5 @@
+export interface SupervisorName {
+  title: string;
+  first: string;
+  last: string;
+}
